Reorder UINT_MAX vs ULONG_MAX checks while defining uint32_t typedef
On Windows, Visual C++, both tests are true.
If ULONG_MAX kicks in first, then uint32_t is unsigned long, while on client side it typically is unsigned int.
In library exports, unsigned long is mangled differently than unsigned int, leading to linker failures with unresolved symbols.

+++ include/liblas/cstdint.hpp
@@ -114,10 +114,10 @@ namespace liblas
 */
 
 #ifndef UINT32_C
-# if ULONG_MAX == 0xffffffff
-    typedef unsigned long   uint32_t;
-# elif UINT_MAX == 0xffffffff
-    typedef unsigned int    uint32_t;
+# if UINT_MAX == 0xffffffff
+    typedef unsigned int  uint32_t;
+# elif ULONG_MAX == 0xffffffff
+    typedef unsigned long uint32_t;
 # else
 #   error defaults not correct; you must hand modify liblas/cstdint.hpp
 # endif
